requirements.txt
pandas
googlesearch-python
streamlit
openpyxl